{"schema":"decodex/commit/1","summary":"Split autonomy objective modules","authority":"manual"}
