+++ apps/decodex/src/autonomy_objective/contract.rs
@@ -1,16 +1,14 @@
 //! Objective Contract payload and lifecycle transitions.
 
+mod accessors;
+mod transitions;
+mod validation_impl;
+
 use serde::{Deserialize, Serialize};
 
-use crate::{
-	autonomy_objective::{
-		lifecycle::{
-			AutonomyObjectiveAcceptance, AutonomyObjectiveRejection, AutonomyObjectiveState,
-			AutonomyObjectiveSupersession,
-		},
-		validation::{self},
-	},
-	prelude::{Result, eyre},
+use crate::autonomy_objective::lifecycle::{
+	AutonomyObjectiveAcceptance, AutonomyObjectiveRejection, AutonomyObjectiveState,
+	AutonomyObjectiveSupersession,
 };
 
 pub(crate) const AUTONOMY_OBJECTIVE_SCHEMA: &str = "decodex.autonomy_objective/1";
@@ -51,325 +49,10 @@ pub(crate) struct AutonomyObjectiveContract {
 	#[serde(skip_serializing_if = "Option::is_none")]
 	supersession: Option<AutonomyObjectiveSupersession>,
 }
-#[allow(dead_code)]
-impl AutonomyObjectiveContract {
-	pub(crate) fn project_id(&self) -> &str {
-		&self.project_id
-	}
-
-	pub(crate) fn id(&self) -> &str {
-		&self.id
-	}
-
-	pub(crate) fn version(&self) -> u64 {
-		self.version
-	}
-
-	pub(crate) fn state(&self) -> AutonomyObjectiveState {
-		self.state
-	}
-
-	pub(crate) fn summary(&self) -> &str {
-		&self.summary
-	}
-
-	pub(crate) fn goals(&self) -> &[String] {
-		&self.goals
-	}
-
-	pub(crate) fn non_goals(&self) -> &[String] {
-		&self.non_goals
-	}
-
-	pub(crate) fn metrics(&self) -> &[String] {
-		&self.metrics
-	}
-
-	pub(crate) fn allowed_surfaces(&self) -> &[String] {
-		&self.allowed_surfaces
-	}
-
-	pub(crate) fn allowed_signal_kinds(&self) -> &[String] {
-		&self.allowed_signal_kinds
-	}
-
-	pub(crate) fn validation_gates(&self) -> &[String] {
-		&self.validation_gates
-	}
-
-	pub(crate) fn review_policy(&self) -> &str {
-		&self.review_policy
-	}
-
-	pub(crate) fn acceptance(&self) -> Option<&AutonomyObjectiveAcceptance> {
-		self.acceptance.as_ref()
-	}
-
-	pub(crate) fn rejection(&self) -> Option<&AutonomyObjectiveRejection> {
-		self.rejection.as_ref()
-	}
-
-	pub(crate) fn supersession(&self) -> Option<&AutonomyObjectiveSupersession> {
-		self.supersession.as_ref()
-	}
-
-	pub(crate) fn validate(&self) -> Result<()> {
-		validation::validate_required("autonomy objective schema", &self.schema)?;
-		validation::validate_required("autonomy objective project_id", &self.project_id)?;
-		validation::validate_required("autonomy objective id", &self.id)?;
-		validation::validate_required("autonomy objective summary", &self.summary)?;
-		validation::validate_required("autonomy objective review_policy", &self.review_policy)?;
-		validation::validate_required("autonomy objective memory_policy", &self.memory_policy)?;
-		validation::validate_required("autonomy objective report_policy", &self.report_policy)?;
-
-		if self.schema != AUTONOMY_OBJECTIVE_SCHEMA {
-			eyre::bail!(
-				"Autonomy objective `{}` has unsupported schema `{}`.",
-				self.id,
-				self.schema
-			);
-		}
-		if self.record_version != AUTONOMY_OBJECTIVE_RECORD_VERSION {
-			eyre::bail!(
-				"Autonomy objective `{}` has unsupported record_version `{}`.",
-				self.id,
-				self.record_version
-			);
-		}
-		if self.version == 0 {
-			eyre::bail!("Autonomy objective `{}` version must be greater than zero.", self.id);
-		}
-
-		validation::validate_string_list("autonomy objective goals", &self.goals)?;
-		validation::validate_string_list("autonomy objective non_goals", &self.non_goals)?;
-		validation::validate_string_list("autonomy objective metrics", &self.metrics)?;
-		validation::validate_string_list(
-			"autonomy objective allowed_surfaces",
-			&self.allowed_surfaces,
-		)?;
-		validation::validate_string_list(
-			"autonomy objective allowed_signal_kinds",
-			&self.allowed_signal_kinds,
-		)?;
-		validation::validate_string_list(
-			"autonomy objective validation_gates",
-			&self.validation_gates,
-		)?;
-
-		self.validate_lifecycle_shape()?;
-		self.validate_lifecycle_records()?;
-
-		Ok(())
-	}
-
-	fn validate_lifecycle_shape(&self) -> Result<()> {
-		match self.state {
-			AutonomyObjectiveState::Draft => self.validate_draft_lifecycle_shape(),
-			AutonomyObjectiveState::Accepted => self.validate_accepted_lifecycle_shape(),
-			AutonomyObjectiveState::Rejected => self.validate_rejected_lifecycle_shape(),
-			AutonomyObjectiveState::Superseded => self.validate_superseded_lifecycle_shape(),
-		}
-	}
-
-	fn validate_lifecycle_records(&self) -> Result<()> {
-		if let Some(acceptance) = &self.acceptance {
-			acceptance.validate()?;
-		}
-		if let Some(rejection) = &self.rejection {
-			rejection.validate()?;
-		}
-		if let Some(supersession) = &self.supersession {
-			supersession.validate()?;
-			self.validate_supersession_target(supersession)?;
-		}
-
-		Ok(())
-	}
-
-	fn validate_draft_lifecycle_shape(&self) -> Result<()> {
-		if self.acceptance.is_some() || self.rejection.is_some() || self.supersession.is_some() {
-			eyre::bail!(
-				"Draft autonomy objective `{}` must not carry lifecycle provenance.",
-				self.id
-			);
-		}
-
-		Ok(())
-	}
-
-	fn validate_accepted_lifecycle_shape(&self) -> Result<()> {
-		if self.acceptance.is_none() {
-			eyre::bail!("Accepted autonomy objective `{}` must include acceptance.", self.id);
-		}
-		if self.rejection.is_some() || self.supersession.is_some() {
-			eyre::bail!(
-				"Accepted autonomy objective `{}` must not carry rejection or supersession.",
-				self.id
-			);
-		}
-
-		self.validate_complete_authority_body()
-	}
-
-	fn validate_rejected_lifecycle_shape(&self) -> Result<()> {
-		if self.rejection.is_none() {
-			eyre::bail!("Rejected autonomy objective `{}` must include rejection.", self.id);
-		}
-		if self.acceptance.is_some() || self.supersession.is_some() {
-			eyre::bail!(
-				"Rejected autonomy objective `{}` must not carry acceptance or supersession.",
-				self.id
-			);
-		}
-
-		Ok(())
-	}
-
-	fn validate_superseded_lifecycle_shape(&self) -> Result<()> {
-		if self.supersession.is_none() {
-			eyre::bail!("Superseded autonomy objective `{}` must include supersession.", self.id);
-		}
-		if self.rejection.is_some() {
-			eyre::bail!("Superseded autonomy objective `{}` must not carry rejection.", self.id);
-		}
-
-		Ok(())
-	}
-
-	fn validate_supersession_target(
-		&self,
-		supersession: &AutonomyObjectiveSupersession,
-	) -> Result<()> {
-		if supersession.superseded_by_objective_id() == self.id
-			&& supersession.superseded_by_version() <= self.version
-		{
-			eyre::bail!(
-				"Autonomy objective `{}` version {} cannot be superseded by same-objective version {}.",
-				self.id,
-				self.version,
-				supersession.superseded_by_version()
-			);
-		}
-
-		Ok(())
-	}
-
-	pub(crate) fn accept(&mut self, acceptance: AutonomyObjectiveAcceptance) -> Result<()> {
-		match self.state {
-			AutonomyObjectiveState::Draft => {},
-			AutonomyObjectiveState::Accepted => {
-				eyre::bail!(
-					"Autonomy objective `{}` version {} is already accepted.",
-					self.id,
-					self.version
-				);
-			},
-			AutonomyObjectiveState::Rejected | AutonomyObjectiveState::Superseded => {
-				eyre::bail!(
-					"Autonomy objective `{}` version {} cannot be accepted from state `{}`.",
-					self.id,
-					self.version,
-					self.state.as_str()
-				);
-			},
-		}
-
-		acceptance.validate()?;
-
-		let mut candidate = self.clone();
-
-		candidate.state = AutonomyObjectiveState::Accepted;
-		candidate.acceptance = Some(acceptance);
-		candidate.rejection = None;
-		candidate.supersession = None;
-
-		candidate.validate()?;
-
-		*self = candidate;
-
-		Ok(())
-	}
-
-	pub(crate) fn reject(&mut self, rejection: AutonomyObjectiveRejection) -> Result<()> {
-		if self.state != AutonomyObjectiveState::Draft {
-			eyre::bail!(
-				"Autonomy objective `{}` version {} can only be rejected from draft state.",
-				self.id,
-				self.version
-			);
-		}
-
-		rejection.validate()?;
-
-		let mut candidate = self.clone();
-
-		candidate.state = AutonomyObjectiveState::Rejected;
-		candidate.acceptance = None;
-		candidate.rejection = Some(rejection);
-		candidate.supersession = None;
-
-		candidate.validate()?;
-
-		*self = candidate;
-
-		Ok(())
-	}
-
-	pub(crate) fn supersede(&mut self, supersession: AutonomyObjectiveSupersession) -> Result<()> {
-		if matches!(
-			self.state,
-			AutonomyObjectiveState::Rejected | AutonomyObjectiveState::Superseded
-		) {
-			eyre::bail!(
-				"Autonomy objective `{}` version {} cannot be superseded from state `{}`.",
-				self.id,
-				self.version,
-				self.state.as_str()
-			);
-		}
-
-		supersession.validate()?;
-
-		let mut candidate = self.clone();
-
-		candidate.state = AutonomyObjectiveState::Superseded;
-		candidate.rejection = None;
-		candidate.supersession = Some(supersession);
-
-		candidate.validate()?;
-
-		*self = candidate;
-
-		Ok(())
-	}
-
-	fn validate_complete_authority_body(&self) -> Result<()> {
-		validation::validate_nonempty_list("accepted autonomy objective goals", &self.goals)?;
-		validation::validate_nonempty_list(
-			"accepted autonomy objective non_goals",
-			&self.non_goals,
-		)?;
-		validation::validate_nonempty_list("accepted autonomy objective metrics", &self.metrics)?;
-		validation::validate_nonempty_list(
-			"accepted autonomy objective allowed_surfaces",
-			&self.allowed_surfaces,
-		)?;
-		validation::validate_nonempty_list(
-			"accepted autonomy objective allowed_signal_kinds",
-			&self.allowed_signal_kinds,
-		)?;
-
-		validation::validate_nonempty_list(
-			"accepted autonomy objective validation_gates",
-			&self.validation_gates,
-		)
-	}
-}
-
-fn autonomy_objective_schema() -> String {
+pub(in crate::autonomy_objective::contract) fn autonomy_objective_schema() -> String {
 	AUTONOMY_OBJECTIVE_SCHEMA.to_owned()
 }
 
-const fn autonomy_objective_record_version() -> u16 {
+pub(in crate::autonomy_objective::contract) const fn autonomy_objective_record_version() -> u16 {
 	AUTONOMY_OBJECTIVE_RECORD_VERSION
 }
